Packing library lib, Update deps, wip
