+++ .babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": ["env"],
+  "presets": ["@babel/preset-env"],
   "plugins": [
     "transform-object-assign"
   ]
