+++ index.js
@@ -1,32 +1,38 @@
-import cookies from './src/cookies';
-import debounce from './src/debounce';
-import events from './src/events';
-import forEach from './src/forEach';
-import safeScrollTop from './src/safeScrollTop';
-import scrollDirection from './src/scrollDirection';
-import scrollTo from './src/scrollTo';
-import touchable from './src/touchable';
+"use strict";
 
-const purejs = {
-  cookies,
-  debounce,
-  events,
-  forEach,
-  safeScrollTop,
-  scrollDirection,
-  scrollTo,
-  touchable
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+
+var _cookies = _interopRequireDefault(require("./lib/cookies"));
+
+var _debounce = _interopRequireDefault(require("./lib/debounce"));
+
+var _events = _interopRequireDefault(require("./lib/events"));
+
+var _forEach = _interopRequireDefault(require("./lib/forEach"));
+
+var _safeScrollTop = _interopRequireDefault(require("./lib/safeScrollTop"));
+
+var _scrollDirection = _interopRequireDefault(require("./lib/scrollDirection"));
+
+var _scrollTo = _interopRequireDefault(require("./lib/scrollTo"));
+
+var _touchable = _interopRequireDefault(require("./lib/touchable"));
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var purejs = {
+  cookies: _cookies.default,
+  debounce: _debounce.default,
+  events: _events.default,
+  forEach: _forEach.default,
+  safeScrollTop: _safeScrollTop.default,
+  scrollDirection: _scrollDirection.default,
+  scrollTo: _scrollTo.default,
+  touchable: _touchable.default
 };
-
-export {
-  cookies,
-  debounce,
-  events,
-  forEach,
-  safeScrollTop,
-  scrollDirection,
-  scrollTo,
-  touchable
-};
-
-export default purejs;
+var _default = purejs;
+exports.default = _default;
+module.exports = exports.default;
