+++ lib/forEach.js
@@ -1,0 +1,18 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = forEach;
+
+/**
+ * https://jsperf.com/for-vs-foreach/37
+ * https://coderwall.com/p/kvzbpa/don-t-use-array-foreach-use-for-instead
+ */
+function forEach(array, callback, scope) {
+  for (var i = 0; i < array.length; i += 1) {
+    callback.call(scope, i, array[i]);
+  }
+}
+
+module.exports = exports.default;
