+++ package.json
@@ -19,33 +19,37 @@
     "node": ">=8.0.0"
   },
   "scripts": {
-    "build": "webpack --env build",
-    "dev": "webpack --env dev",
+    "lint": "eslint src test",
+    "lint:fix": "eslint --fix src test",
+    "clean:node": "rm -rf index.js lib",
+    "clean:browser": "rm -rf validator*.js",
+    "clean": "npm run clean:node && npm run clean:browser",
+    "minify": "uglifyjs validator.js -o validator.min.js  --compress --mangle --comments /Copyright/",
+    "build:browser": "webpack",
+    "build:node": "babel src -d . --presets @babel/preset-env --plugins add-module-exports",
+    "build": "npm run build:browser && npm run build:node",
+    "pretest": "npm run lint && npm run build",
     "test": "karma start",
     "updeps": "ncu"
   },
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.3",
-    "babel-eslint": "8.2.5",
-    "babel-loader": "7.1.5",
-    "babel-plugin-add-module-exports": "0.2.1",
+    "@babel/cli": "^7.1.0",
+    "@babel/core": "7.1.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/register": "^7.0.0",
+    "babel-eslint": "9.0.0",
+    "babel-loader": "8.0.2",
+    "babel-plugin-add-module-exports": "1.0.0",
     "babel-plugin-transform-object-assign": "6.22.0",
-    "babel-preset-env": "1.7.0",
     "chai": "4.1.2",
-    "eslint": "5.1.0",
-    "eslint-config-airbnb": "17.0.0",
-    "eslint-config-prettier": "2.9.0",
-    "eslint-config-react": "1.1.7",
-    "eslint-loader": "2.0.0",
-    "eslint-plugin-flow": "2.29.1",
-    "eslint-plugin-flowtype": "2.50.0",
-    "eslint-plugin-import": "2.13.0",
-    "eslint-plugin-jsx-a11y": "6.1.0",
+    "eslint": "5.6.0",
+    "eslint-config-airbnb-base": "13.1.0",
+    "eslint-config-prettier": "3.0.1",
+    "eslint-loader": "2.1.0",
+    "eslint-plugin-import": "2.14.0",
     "eslint-plugin-prettier": "2.6.2",
-    "eslint-plugin-react": "7.10.0",
     "istanbul-instrumenter-loader": "3.0.1",
-    "karma": "2.0.4",
+    "karma": "3.0.0",
     "karma-babel-preprocessor": "7.0.0",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.2.0",
@@ -56,11 +60,11 @@
     "karma-sinon-chai": "2.0.2",
     "mocha": "5.2.0",
     "npm-check-updates": "2.14.2",
-    "sinon": "6.1.3",
+    "sinon": "6.3.4",
     "sinon-chai": "3.2.0",
-    "uglifyjs-webpack-plugin": "1.2.7",
-    "webpack": "4.16.0",
-    "webpack-cli": "3.0.8",
-    "yargs": "12.0.1"
+    "uglifyjs-webpack-plugin": "2.0.1",
+    "webpack": "4.19.1",
+    "webpack-cli": "3.1.0",
+    "yargs": "12.0.2"
   }
 }
