+++ src/index.js
@@ -1,0 +1,21 @@
+import cookies from './lib/cookies';
+import debounce from './lib/debounce';
+import events from './lib/events';
+import forEach from './lib/forEach';
+import safeScrollTop from './lib/safeScrollTop';
+import scrollDirection from './lib/scrollDirection';
+import scrollTo from './lib/scrollTo';
+import touchable from './lib/touchable';
+
+const purejs = {
+  cookies,
+  debounce,
+  events,
+  forEach,
+  safeScrollTop,
+  scrollDirection,
+  scrollTo,
+  touchable
+};
+
+export default purejs;
