+++ src/lib/safeScrollTop.js
