+++ src/lib/scrollDirection.js
