+++ webpack.config.babel.js
@@ -5,37 +5,28 @@ import { argv } from 'yargs';
 const { env } = argv;
 const libraryName = 'purejs';
 const plugins = [];
-const minimizer = [];
-
-let suffix = '';
-let mode = 'development';
-
-if (env === 'build') {
-  minimizer.push(
-    new UglifyJsPlugin({
-      cache: false,
-      parallel: true,
-      uglifyOptions: {
-        compress: true,
-        ecma: 6,
-        mangle: false
-      },
-      sourceMap: false
-    })
-  );
-  mode = 'production';
-  suffix = '.min';
-}
+const minimizer = [
+  new UglifyJsPlugin({
+    cache: false,
+    parallel: true,
+    uglifyOptions: {
+      compress: true,
+      ecma: 6,
+      mangle: false
+    },
+    sourceMap: false
+  })
+];
 
 const config = {
   entry: {
-    purejs: `${__dirname}/index.js`
+    purejs: `${__dirname}/src/index.js`
   },
-  mode,
+  mode: 'production',
   devtool: 'source-map',
   output: {
-    path: `${__dirname}/lib`,
-    filename: `[name]${suffix}.js`,
+    path: `${__dirname}`,
+    filename: `[name].min.js`,
     library: libraryName,
     libraryTarget: 'umd',
     umdNamedDefine: true
@@ -43,12 +34,12 @@ const config = {
   module: {
     rules: [
       {
-        test: /(\.jsx|\.js)$/,
+        test: /(\.js)$/,
         loader: 'babel-loader',
         exclude: /(node_modules)/
       },
       {
-        test: /(\.jsx|\.js)$/,
+        test: /(\.js)$/,
         loader: 'eslint-loader',
         exclude: /node_modules/
       }
